Migrating to Spring Boot

+++ src/main/java/com/epam/esm/service/TagService.java
@@ -11,6 +11,7 @@ import com.epam.esm.exceptions.ErrorCode;
 import com.epam.esm.model.Tag;
 import com.epam.esm.repository.TagRepository;
 import java.util.Optional;
+import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
@@ -30,6 +31,14 @@ public class TagService {
     }
 
 
+    /**
+     *
+     * @param tagName String value that can be empty or null
+     * @return cases:
+     * empty/null: Not valid request - TAG_BAD_REQUEST
+     * already existing name: Tag already exists - TAG_ALREADY_FOUND
+     * successfully saved: return TagResponseDTO
+     */
     public TagResponseDTO saveTag(String tagName) {
 
         if (tagName == null || tagName.isEmpty()) {
@@ -38,7 +47,7 @@ public class TagService {
 
         tagRepository.findByName(tagName)
             .ifPresent(tag -> {
-                throw new CustomizedException(TAG_ALREADY_EXISTS.formatted(tag.getId()), ErrorCode.TAG_BAD_REQUEST);
+                throw new CustomizedException(TAG_ALREADY_EXISTS.formatted(tag.getId()), ErrorCode.TAG_ALREADY_FOUND);
             });
 
         Tag tag = new Tag();
@@ -47,12 +56,20 @@ public class TagService {
         try {
             Tag savedTag = tagRepository.save(tag);
             return convertTagToTagReturnDTO(savedTag);
-        }catch (DataAccessException ex){
+        } catch (DataAccessException ex){
             throw new CustomizedException(TAG_CANNOT_BE_SAVED, ErrorCode.TAG_DATABASE_ERROR, ex);
         }
     }
 
-    public Page<TagResponseDTO> getAllTags(Pageable pageable) {
+    /**
+     *
+     * @param pageable cannot be null or empty
+     * @return cases:
+     * Page type of TagResponseDTO
+     * Fail fetching tags - TAG_DATABASE_ERROR
+     * Other error - TAG_DATABASE_ERROR
+     */
+    public Page<TagResponseDTO> getAllTags(@NonNull Pageable pageable) {
         try {
             return tagRepository.findAll(pageable)
                 .map(this::convertTagToTagReturnDTO);
@@ -65,7 +82,14 @@ public class TagService {
         }
     }
 
-    public TagResponseDTO getMostPopularTag(){
+    /**
+     *
+     * @return cases:
+     * most used tag exists, return TagResponseDTO
+     * there is no tag, return null
+     * error fetching tag - TAG_DATABASE_ERROR
+     */
+    public TagResponseDTO getMostUsedTag(){
         try {
             Optional<Long> tagId = tagRepository.findMostUsedTagOfUserWithHighestTotalCostOfOrders();
             if(tagId.isPresent()) {
@@ -74,15 +98,29 @@ public class TagService {
             }
             else return null;
         } catch (DataAccessException ex) {
-            throw new CustomizedException("Database error during getting tag", ErrorCode.TAG_DATABASE_ERROR, ex);
+            throw new CustomizedException("Database error while fetching tag", ErrorCode.TAG_DATABASE_ERROR, ex);
         }
     }
 
 
+    /**
+     *
+     * @param tagId should not be less or equal than zero
+     * cases:
+     * tagId is less than zero: NOT_VALID_TAG_REQUEST, TAG_BAD_REQUEST
+     * tagId does not exists: TAG_ID_NOT_FOUND, TAG_NOT_FOUND
+     * tag successfully deleted
+     * error deleting tag, TAG_DATABASE_ERROR
+     */
     public void deleteTag(long tagId) {
+        if (tagId <= 0) {
+            throw new CustomizedException(NOT_VALID_TAG_REQUEST, ErrorCode.TAG_BAD_REQUEST);
+        }
+
         if (!tagRepository.existsById(tagId)) {
             throw new CustomizedException(TAG_ID_NOT_FOUND.formatted(tagId), ErrorCode.TAG_NOT_FOUND);
         }
+
         try {
             tagRepository.deleteById(tagId);
         } catch (DataAccessException ex) {
